% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19Characterization.R
\docType{package}
\name{Covid19Characterization-package}
\alias{Covid19Characterization}
\alias{Covid19Characterization-package}
\title{Covid19Characterization: Characterization of COVID-19 and Influenza patients}
\description{
A package for characterizing COVID-19 patients with and without various features.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\keyword{internal}
