% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{createBulkCharacteristics}
\alias{createBulkCharacteristics}
\title{Create cohort characteristics in bulk}
\usage{
createBulkCharacteristics(
  connection,
  oracleTempSchema,
  cohortIds,
  cdmDatabaseSchema,
  cohortDatabaseSchema,
  cohortTable
)
}
\description{
This function will perform the same actions as the main RunStudy.R::runStudy()
function but in a single SQL operation.
}
