% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\name{writeBulkCharacteristics}
\alias{writeBulkCharacteristics}
\title{Write cohort characteristics in bulk to the file system}
\usage{
writeBulkCharacteristics(
  connection,
  oracleTempSchema,
  counts,
  minCellCount,
  databaseId,
  exportFolder
)
}
\description{
This function will retrieve the results from the temp tables created
in createBulkCharacteristics
}
