% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid19CharacterizationCharybdis.R
\docType{package}
\name{Covid19CharacterizationCharybdis-package}
\alias{Covid19CharacterizationCharybdis}
\alias{Covid19CharacterizationCharybdis-package}
\title{Covid19CharacterizationCharybdis: Characterization of COVID-19 and Influenza patients}
\description{
A package for characterizing COVID-19 patients with and without various features.
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\keyword{internal}
